+++ src/config.rs
@@ -11,7 +11,7 @@ pub struct Config {
 
 #[derive(Deserialize, Serialize)]
 pub struct WatchedDevice {
-    pub product: String,
+    pub properties: HashMap<String, String>,
     pub on_plugged: String,
     pub on_unplugged: String,
 }
@@ -22,19 +22,25 @@ impl Config {
         let mut buffer = String::new();
         file.read_to_string(&mut buffer)?;
 
-        let config: Config = toml::from_str(&buffer).unwrap();
-
-        Ok(config)
+        match toml::from_str(&buffer) {
+            Ok(config) => Ok(config),
+            Err(e) => {
+                eprintln!("{}", e);
+                std::process::exit(1);
+            }
+        }
     }
 }
 
 impl Default for Config {
     fn default() -> Config {
+        let mut properties = HashMap::new();
+        properties.insert("product".to_owned(), "3/feed/1307/111".to_owned());
         let mut devices = HashMap::new();
         devices.insert(
             "ergodox".to_owned(),
             WatchedDevice {
-                product: "3/feed/1307/111".to_owned(),
+                properties,
                 on_plugged: "setxkbmap us".to_owned(),
                 on_unplugged: "setxkbmap fr oss -option ctrl:nocaps".to_owned(),
             },