+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Add option to list current devices
 - Add option to describe device properties
+- Allow to match device based on property set
 
 ## [0.2.0] - 2018-09-03
 ### Fixed